requirements.txt
astroid==1.4.6
asyncio==3.4.3
click==6.6
colorama==0.3.7
coverage==4.1
coveralls==1.1
docopt==0.6.2
lazy-object-proxy==1.2.2
ordereddict==1.1
pluggy==0.3.1
py==1.4.31
pylint==1.5.6
pytest==2.9.2
pytest-cov==2.3.0
python-dotenv==0.5.1
requests==2.10.0
six==1.10.0
virtualenv==15.0.2
wrapt==1.10.8
